API Reorganise file structure
