+++ _config/sitewidecontent-report.yml
@@ -3,6 +3,6 @@ Name: reports-sitewidecontenttaxonomy
 Only:
   moduleexists: silverstripe/taxonomy
 ---
-SilverStripe\Reports\SiteWideContentReport\SitewideContentReport:
+SilverStripe\Reports\SitewideContentReport\Reports\SitewideContentReport:
   extensions:
-    - SilverStripe\Reports\SiteWideContentReport\Model\SitewideContentTaxonomy
+    - SilverStripe\Reports\SiteWideContentReport\Extensions\SitewideContentTaxonomy
